Refactor Config handling and imports
- Replaced Config class with an interface in types directory
- Updated imports across core and custom-views files
